refactor(cache): split into directory + address review findings
Addresses thermo-nuclear and caveman review feedback on the prior commit:

Structure
- Split `src/modules/cache.ts` (913 LOC, four concerns) into a directory:
  - `cache/types.ts`     — public + internal types, column list, prefixes
  - `cache/db.ts`        — connection lifecycle, mirror, upsert/delete
  - `cache/read.ts`      — sync read API (mirror-only)
  - `cache/write.ts`     — async write API (SQLite first, then mirror) +
                            shared `setExtraConfirmedMatch` helper
  - `cache/migration.ts` — one-shot Extra→SQLite migration + orphan GC
  - `cache/index.ts`     — re-exports preserving the import path
                            `"../modules/cache"` for all existing callers
- Largest file is now 329 LOC (migration.ts); every other module is
  under 220 LOC. Migration can be deleted as a unit at v2.x without
  disturbing the runtime cache.

Correctness fixes
- `deriveCitationMetrics` helper in `write.ts`: the `cited_by_count > 0`
  guard now applies uniformly to fwci/percentile/isTop1/isTop10. The
  prior code applied it only to fwci+percentile, so a work with 0
  citations but a non-null `citation_normalized_percentile` could land
  inconsistent `is_top_*_percent` values alongside a null percentile.
- `verifyParseRoundTrip` now does sorted-multiset comparison instead of
  Set comparison. Set semantics silently collapsed duplicate lines, so
  the round-trip check could approve a migration that loses a repeated
  line (real-world case: BibTeX import scripts leave duplicate `PMID:`
  lines). Multiset comparison preserves duplicate counts.
- `garbageCollectOrphans` now unions item keys across **every** library
  via `Zotero.Libraries.getAll()`. The prior version only checked the
  user library, so group-library items written to SQLite via the
  runtime write path would be purged as "orphans" on every startup.
- `garbageCollectOrphans` is now rate-limited by
  `ORPHAN_GC_MIN_INTERVAL_MS` (7 days) and accepts `{ force: true }`
  for explicit-rebuild paths. Eliminates the per-startup
  full-library-scan cost for users with stable libraries.

Type tightening
- New `DbBool = 0 | 1 | null` type for SQLite-stored boolean columns,
  replacing `number | null` with a comment. The compiler now catches
  accidental assignment of values like `2` or `-1`.
- Stripped `executeTransaction` and `tableExists` from the
  `DBConnection` typing — neither is used anywhere in `src/`, and
  advertising them in the type encouraged readers to believe an
  atomic-batch path existed.
- Stripped the unused `Zotero.DB` declaration. Only
  `new Zotero.DBConnection(...)` is used.
- Added `_ZoteroTypes.Library` and `Libraries.getAll()` to typings to
  back the cross-library GC.

Code quality
- `confirmTitleMatch` now logs and no-ops when called with no pending
  or existing OpenAlex ID (prior: silently persisted a malformed row
  with `confirmed_open_alex_id: null`). Makes caller invariant
  violations visible.
- `writeConfirmedMatchToExtra` and the migration's confirmed-match
  re-emit now share `setExtraConfirmedMatch(lines, openAlexId)` — one
  copy of the filter+optional-push rule, used by both call sites.
- `cacheReady()` predicate exported so future callers can guard writes
  on a successful init.
- Magic number `200` in orphan GC chunk size moved to
  `ORPHAN_GC_CHUNK_SIZE` in `constants.ts`.

Tests
- 185 passing. Updated GC tests to pass `{ force: true }` and to mock
  `Zotero.Libraries.getAll`. No test logic changed.

Build
- Typecheck clean. Lint 0 errors. Prettier clean. XPI 72.8 KB.
