+++ src/constants.ts
@@ -54,6 +54,10 @@ export const MAX_ABSTRACT_LENGTH = 100_000;
 export const SHOW_PROGRESS_UI_THRESHOLD = 500;
 /** Update progress UI every N items during migration. */
 export const MIGRATION_PROGRESS_TICK = 50;
+/** Max item_keys per `DELETE … WHERE item_key IN (…)` chunk during orphan GC. */
+export const ORPHAN_GC_CHUNK_SIZE = 200;
+/** Minimum interval between orphan-GC sweeps at startup. */
+export const ORPHAN_GC_MIN_INTERVAL_MS = 7 * 24 * 60 * 60 * 1000; // 7 days
 
 // ── Title-based metadata matching ──
 /** Score threshold for high-confidence title match (data shown with ~ prefix). */
